+++ src/components/Lists/index.js
@@ -3,32 +3,32 @@ import { Link } from 'react-router-dom';
 
 import './Lists.css';
 
-export const Track = ({ val }) => {
-  const song = val.trackName.toLowerCase().replace(/ /g, '+');
-  const album = val.collectionName.toLowerCase().replace(/ /g, '+');
+export const Track = ({ value }) => {
+  const song = value.trackName.toLowerCase().replace(/ /g, '+');
+  const album = value.collectionName.toLowerCase().replace(/ /g, '+');
 
   return (
     <div className="track">
       <div className="track__primary">
         <img
           className="track__img"
-          src={val.artworkUrl60}
-          alt={val.trackName}
+          src={value.artworkUrl60}
+          alt={value.trackName}
         />
         <audio controls className="player" preload="false">
-          <source src={val.previewUrl} />
+          <source src={value.previewUrl} />
         </audio>
       </div>
 
       <div className="track__secondary">
-        <Link to={`/song/${song}/${val.trackId}`}>
-          <p className="song-title">{val.trackName}</p>
+        <Link to={`/song/${song}/${value.trackId}`}>
+          <p className="song-title">{value.trackName}</p>
         </Link>
-        <div className={val.trackExplicitness} />
+        <div className={value.trackExplicitness} />
 
-        <Link to={`/album/${album}/${val.collectionId}`}>
+        <Link to={`/album/${album}/${value.collectionId}`}>
           <p className="album">
-            {val.collectionName} &bull; {val.releaseDate.substring(0, 4)}
+            {value.collectionName} &bull; {value.releaseDate.substring(0, 4)}
           </p>
         </Link>
       </div>
@@ -36,30 +36,54 @@ export const Track = ({ val }) => {
   );
 };
 
-export const Collection = ({ val }) => {
-  const artist = val.artistName.toLowerCase().replace(/ /g, '+');
-  const album = val.collectionName.toLowerCase().replace(/ /g, '+');
+export const Video = ({ value }) => {
+  const video = value.trackName.toLowerCase().replace(/ /g, '+');
+  //const album = value.collectionName.toLowerCase().replace(/ /g, '+');
+
+  return (
+    <div className="track">
+      <div className="track__primary">
+        <img
+          className="track__img"
+          src={value.artworkUrl60}
+          alt={value.trackName}
+        />
+      </div>
+
+      <div className="track__secondary">
+        <Link to={`/music-video/${video}/${value.trackId}`}>
+          <p className="song-title">{value.trackName}</p>
+        </Link>
+        <div className={value.trackExplicitness} />
+      </div>
+    </div>
+  );
+};
+
+export const Collection = ({ value }) => {
+  const artist = value.artistName.toLowerCase().replace(/ /g, '+');
+  const album = value.collectionName.toLowerCase().replace(/ /g, '+');
 
   return (
     <div className="collection">
       <div className="collection__primary">
         <img
           className="collection__img"
-          src={val.artworkUrl100}
-          alt={val.trackName}
+          src={value.artworkUrl100}
+          alt={value.trackName}
         />
       </div>
 
       <div className="collection__secondary">
-        <Link to={`/album/${album}/${val.collectionId}`}>
-          <p className="album">{val.collectionName}</p>
+        <Link to={`/album/${album}/${value.collectionId}`}>
+          <p className="album">{value.collectionName}</p>
         </Link>
 
-        <Link to={`/artist/${artist}/${val.artistId}`}>
-          <p className="author">{val.artistName}</p>
+        <Link to={`/artist/${artist}/${value.artistId}`}>
+          <p className="author">{value.artistName}</p>
         </Link>
 
-        <div className={val.trackExplicitness} />
+        <div className={value.trackExplicitness} />
       </div>
     </div>
   );