+++ src/components/Menu/index.js
@@ -5,7 +5,7 @@ import SearchForm from '../SearchForm';
 import './Menu.css';
 
 export default class Menu extends Component {
-  getValue = val => this.props.dataSearch(val);
+  getValue = value => this.props.dataSearch(value);
 
   render() {
     return (