+++ src/containers/Search/index.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
-import { Track, Collection } from '../../components/Lists';
+import { Track, Video, Collection } from '../../components/Lists';
 import Spinner from '../../components/Spinner';
 import { musicSearch } from '../../api';
 
 export default class Search extends Component {
-  state = { data: false, tracks: null, collections: null };
+  state = { data: false, tracks: null, videos: null, collections: null };
 
   componentDidMount = () => this.results(this.props.match.params.query);
   //getSnapshotBeforeUpdate = value => this.results(value.match.params.query);
@@ -18,16 +18,19 @@ export default class Search extends Component {
 
   parts() {
     const tracks = [];
+    const videos = [];
     const collections = [];
 
     for (const type of this.state.data) {
-      if (type.wrapperType === 'track') tracks.push(type);
+      if (type.kind === 'song') tracks.push(type);
+      else if (type.kind === 'music-video') videos.push(type);
       else if (type.wrapperType === 'collection') collections.push(type);
     }
 
     this.setState({
-      tracks: tracks.map(val => <Track val={val} />),
-      collections: collections.map(val => <Collection val={val} />)
+      tracks: tracks.map(value => <Track value={value} />),
+      videos: videos.map(value => <Video value={value} />),
+      collections: collections.map(value => <Collection value={value} />)
     });
   }
 
@@ -41,6 +44,10 @@ export default class Search extends Component {
         <div className="collections">
           <h3 className="collections__title">Albums</h3>
           <div className="collections__container">{this.state.collections}</div>
+        </div>
+        <div className="videos">
+          <h3 className="videos__title">Videos</h3>
+          <div className="videos__container">{this.state.videos}</div>
         </div>
       </React.Fragment>
     ) : (